backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.49
asyncpg==0.31.0
alembic==1.14.0

# Settings
pydantic-settings==2.6.1
pydantic[email]>=2.13.4

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12

# AI
groq==0.13.0

# PDF parsing
PyMuPDF==1.27.2.3

# HTTP client (for AI calls)
httpx==0.27.2

# PDF generation + charts
reportlab==4.5.0
matplotlib==3.10.9
Pillow==12.2.0
